refactor: rename radio component
