+++ src/files/modals/publish-modal/PublishModal.js
@@ -6,7 +6,7 @@ import Button from '../../../components/button/button.tsx'
 import { Modal, ModalActions, ModalBody } from '../../../components/modal/modal'
 import Icon from '../../../icons/StrokeSpeaker.js'
 import { connect } from 'redux-bundler-react'
-import Radio from '../../../components/radio/radio-component.js'
+import Radio from '../../../components/radio/radio.js'
 import ProgressBar from '../../../components/progress-bar/ProgressBar.js'
 import GlyphCopy from '../../../icons/GlyphCopy.js'
 import GlyphTick from '../../../icons/GlyphTick.js'
